feat(emeisOptions): passing options as a function
- this enable to pass options for meta-fields of type choice as static list or as a function
