--- addon/components/meta-fields.hbs
@@ -1,30 +1,0 @@
-{{#each @fields as |field|}}
-  {{#if (eval-meta field.visible @model)}}
-    <EditForm::Element @label={{optional-translate field.label}}>
-      {{#if (eq field.type "choice")}}
-        <PowerSelect
-          @disabled={{eval-meta field.readOnly @model}}
-          @options={{field.options}}
-          @selected={{find-by "value" (get @model.metainfo field.slug) field.options}}
-          @onChange={{fn this.updateMetaField field @model}}
-          @placeholder={{t field.label}}
-          @allowClear={{true}}
-          @renderInPlace={{true}}
-          as |option|
-        >
-          {{optional-translate option.label}}
-        </PowerSelect>
-      {{/if}}
-      {{#if (eq field.type "text")}}
-        <input
-          data-test-meta-field-text={{field.slug}}
-          class="uk-input"
-          type="text"
-          value={{get @model.metainfo field.slug}}
-          disabled={{eval-meta field.readOnly @model}}
-          {{on "input" (fn this.updateMetaField field @model)}}
-        >
-      {{/if}}
-    </EditForm::Element>
-  {{/if}}
-{{/each}}
