+++ app/components/meta-field.js
@@ -1,0 +1,1 @@
+export { default } from "ember-emeis/components/meta-field";
