--- app/components/meta-fields.js
@@ -1,1 +1,0 @@
-export { default } from "ember-emeis/components/meta-fields";
