--- app/helpers/eval-meta.js
@@ -1,1 +1,0 @@
-export { default } from "ember-emeis/helpers/eval-meta";
